+++ CHANGES.md
@@ -1,3 +1,7 @@
+# 0.3.0
+* Support creating a reader pipe from `Input_channel`.
+* Support creating a writer pipe from `Output_channel`.
+
 # 0.2.0
 
 * Support deferred's in `Input_channel.read_one_chunk_at_a_time`